feat: add app config for invitations, notification center, bidali (#113)
### Description

As the title. These are features that most apps will not want to enable
(e.g. cKash).

The Bidali integration is also baked into the dapps list (the url
provided is the deep link rather than the website url). I think keeping
the deep link flow to Bidali alive seems low risk, as then we don't need
to add special logic to handle Bidali for apps that will use the dapp
list.

### Test plan

n/a

### Related issues

Related to RET-1309

### Backwards compatibility

Yes

### Network scalability

N/A