feat: pass prepared transaction to the SendConfirmation screen with navigate (#102)
### Description
This PR implements passing prepared transaction from `SendEnterAmount`
to `SendConfirmation` with `navigate` to prevent unnecessary call to
prepare transaction again. The call still can be called when navigating
to `SendConfirmation` with `Screen.SendConfirmationFromExternal`.

### Test plan
It doesn't load anything and doesn't show loading spinners when
navigating to the SendConfirmation


https://github.com/user-attachments/assets/dd63ede5-fef9-42a8-bc65-5ed15266bfb2



### Related issues

Relates to
[RET-1300](https://linear.app/valora/issue/RET-1300/%5Bwallet%5D-refactor-sendconfirmation-to-optionally-re-run-prepared)

### Backwards compatibility
Yes 

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)