feat: add config for phone verification (#118)
### Description

As the title.

Note that the profile submenu is bypassed entirely if there is no phone
verification. Checked with Kayla
[here](https://valora-app.slack.com/archives/C0684TXDR3K/p1740747988181399)

### Test plan

The entry points to phone verification are removed.



https://github.com/user-attachments/assets/5d0f74f1-0157-4728-8c55-c01f260a4849


### Related issues

- Fixes RET-1309

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)