feat: add InfoBottomSheet (#122)
### Description
As per the title. I need this as a part of my work on refactoring Earn
deposit confirmation flow to use new review screen. Moved this out of
ReviewTransaction component as it is gonna be used for all info sheets
on send/earn/swap. This is a very simple abstraction with button always
sticking to the bottom and content being separated with spacing away
from it.

### Test plan
No tests needed, this is a really simple component the only testing of
which would involve just checking if all the parts rendered, which is
unnecessary here.

### Related issues

Relates to
[RET-1302](https://linear.app/valora/issue/RET-1302/%5Bwallet%5D-add-new-reviewtransaction-component-to-earndepositbottomsheet)

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)