feat: move FeeInfoBottomSheet to a reusable component (#124)
### Description
As per the title.
- `FeeInfoBottomSheet` needs to be used across all the flows
consistently and also earn deposit flow already contained a `TODO`
comment stating a need to reuse `src/swap/FeeInfoBottomSheet.tsx`
- this is the same `FeeInfoBottomSheet` from `src/swap` with the only
changes are each fee label and `FeeAmount` being replaced with a new
`ReviewDetailsItem` for consistency
- removed `fetchingSwapQuote` and the corresponding test since after
migrating to `SwapScreenV2` we don't show fees breakdown until the quote
is loaded
- moved most of the translations to the top level as most of them seem
to be ambiguous enough to be on a top level. Others are moved to a new
contextual key
- adjusted spacing between items to be more on par with the updated
design

### Test plan
Reused existing tests for the FeeInfoBottomSheet and removed the
unnecessary ones.

### Related issues

Relates to
[RET-1302](https://linear.app/valora/issue/RET-1302/%5Bwallet%5D-add-new-reviewtransaction-component-to-earndepositbottomsheet)

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)