refactor(tokens): remove network id param from selectors (#132)
### Description

Now that we've dropped per feature network id configuration, the token
selectors no longer need to take networkIds to filter out tokens.

The primary change for this PR is in `src/tokens/selectors.ts`.
Specifically the top level selector now filters tokens using the
`getSupportedNetworkIds` util instead of the networkIds param.

Others are just changes in selector usages to drop the param. Some
usages were only passing a single network Id, but those were then just
extracting required tokens by tokenId, so we don't need the filtering by
network Id there. Others usages where filtering was required, it is just
done locally after getting all tokens.

More cleanup, like hooks which now just wrap a useSelector, can be done
in a follow up

### Test plan

CI

### Related issues

N/A

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)