feat: Add (de)serialization for PossibleTransactionResult and apply it to SendConfirmation (#137)
### Description
Based on
[this](https://github.com/divvi-xyz/divvi-mobile/pull/102#discussion_r1977255177)
and
[this](https://github.com/divvi-xyz/divvi-mobile/pull/136#discussion_r1985851448)
review comments it appeared valuable to have a way to
serialize/deserialize possible prepared transaction results in order to
be able to pass those as a params of the screen using
`react-navigation`. In order to achieve this – `TokenBalance` also
needed to have a (de)serialization mechanism as
`PreparedTransactionsPossible` has a `feeCurrency` property which is of
that type. It is used by `getFeeCurrencyAndAmounts`.

This appeared needed after working on `SendConfirmation` and new
`EarnDepositConfirmationScreen` screens.

This PR adds:
- serialization/deserialization functions for TokenBalance
- serialization/deserialization functions for
PreparedTransactionsPossible
- applying this logic to `SendConfirmation` screen.
### Test plan
- added tests for serialization/deserialization of token balances
- added tests for serialization/deserialization of possible prepared
transaction results

### Related issues

Relates to
[ENG-184](https://linear.app/valora/issue/ENG-184/%5Bwallet%5D-add-new-reviewtransaction-component-to-earndepositbottomsheet)

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)