refactor: get crowdin lang from getContent (#151)
### Description

Follow up from #142 - with @jeanregisser we were questioning if there
was any way to avoid a custom language mapping that we need to maintain.
The `getContent` api seems like a good start, with the tradeoff that we
need to make some assumption about the content value string format. It
does not add an extra api request since the manifest is cached under the
hood in Crowdin and is used by `getStringsByLocale` anyway.

Along the way of testing, I noticed that we are not invoking the OTA
saga on app start anymore. Idk when this broke, but I added an explicit
spawn in the i18n saga.

### Test plan

Tested that on app start and on app change language, OTA translations
are fetched correctly and applied over the bundled translations when
there is a difference.

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)