feat: use new `EarnDepositConfirmationScreen` (#159)
### Description
As per the title.
- removed `EarnDepositBottomSheet.tsx` and
`EarnDepositBottomSheet.test.tsx`
- added navigation to `Screens.EarnDepositConfirmationScreen` when
continue is pressed on the earn deposit flow

### Test plan
Adjusted existing tests on `EarnEnterAmount.test.tsx` to expect
navigation with params.

Full flow:


https://github.com/user-attachments/assets/fafb19b2-126d-4653-91d1-5f2d3c8dc316


### Related issues

Relates to
[ENG-184](https://linear.app/valora/issue/ENG-184/%5Bwallet%5D-add-new-reviewtransaction-component-to-earndepositbottomsheet)

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)