chore: move `useNetworkFee` for reusability (#160)
### Description
As I was applying new review components to the existing
`EarnEnterAmount` screen for consistency, I noticed that formatting of
the network fee is identical on all of the following screens:
- `EarnDepositConfirmationScreen` (already uses new design)
- `EarnEnterAmount` (WIP applying new design to transaction details)
- `EarnConfirmationScreen` (will migrate to new design as a next task)
All of those screens are supposed to also have `FeeInfoBottomSheet`
which requires network fee data as well so it seems beneficial to have
`useNetworkFee` as a reusable earn-related hook.

### Test plan
Moved existing test to a new file.

### Related issues
N/A

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)