feat: enable crowdin translations (#164)
### Description

This PR brings back the Crowdin translation sync. The sync config has
been updated in
[Crowdin](https://valorainc.crowdin.com/u/projects/2/integrations/system/github),
as well as the crowdin.yml in this repo.

### Test plan

The divvi-bot-owner will now submit PRs (e.g. #166 )

### Related issues

- Related to ENG-196

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)