feat!: don't expose testnets for now (#180)
### Description

The framework currently only works for mainnet/production networks.
So it's preferable not to expose testnets for now.

### Test plan

- Tests pass

### Related issues

N/A

### Backwards compatibility

No, but no (known) divvi apps were using testnets.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)