fix: estimateFeesPerGas without getGasPriceMinimum on Celo (#188)
### Description

This updates `estimateFeesPerGas` to use Celo's fee estimation
implementation in viem without relying on `getGasPriceMinimum` which
can't be used anymore following the L2 hardfork.

The added benefit is that it now follows the
[baseFeeMultiplier](https://rc.viem.sh/docs/clients/chains.html#fees-basefeemultiplier)
applied by viem and should be more future proof.

The `baseFeePerGas` is captured in a hacky way, but unit tests are
present to ensure it will keep working as expected.

Viem was updated so we get the latest implementation for
`estimateFeesPerGas` on Celo L2.

### Test plan

- Updated tests
- Manually tested the network fees are correct with different fee
currencies.

### Related issues

- N/A

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)