feat: extend functionality of `total-token-amount` version of `ReviewDetailsItem` component (#175)
### Description
While working on the new earn withdrawal confirmation screen as per the
design changes there appeared to be two issues with the current
implementation of `ReviewDetailsItem` component which is used to show
total amount for the transaction:
- there is no way to deduct fee from the amount instead of adding it to
the amount
- there is no way to add more token amounts to the total (e.g. to
include claimed rewards to the total).
For this case there was a need to change the implementation of the
`total-token-amount` type of `ReviewDetailsItem` component to allow
these calculations.

Initially I considered to just extend the props to allow all the
possible calculation variations but there are too many amounts that
could be added:
- swap app fee
- cross chain fee
- multiple claimed rewards

So it felt a lot more natural to lean to a more automated approach using
dynamic array of amounts.

There is no memoization in `ReviewDetailsItem` for the following
reasons:
- the `amounts` array is expected to always be tiny (between 1-5 items
in 99% of cases) so the overhead of re-running the loop functions can be
considered negligible
- memoization would require the consumer components to always memoize
the passed array, which we definitely can forget to do (or forget to
make it type-safe).

There is also no memoization in the consumer components as using
`useMemo` outside of the scope of JSX (meaning above the `return`
statement) to declare the array of amounts would require the type for
the array to be exported from `ReviewTransaction.tsx` file in order to
still keep the type-safety. It feels unnecessary and redundant (in terms
of extra boilerplate) considering the array is always gonna be small and
we can get the type-safety simply from defining the array right on the
JSX level.

I've added some extra comments to the `ReviewDetailsItemTotalValue`
component to better explain the logic of the component so it is more
clear what the rendering conditions for each format are.

<!-- A few sentences describing the overall effects and goals of the
pull request's commits.
What is the current behavior, and what is the updated/expected behavior
with this PR? -->

### Test plan
Adjusted existing tests to the new props.

### Related issues

- Relates to
[ENG-223](https://linear.app/valora/issue/ENG-223/%5Bwallet%5D-add-new-reviewtransaction-component-to-earnconfirmationscreen)

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)