feat: use new `EarnWithdrawConfirmationScreen` (#187)
### Description
This PR removes an old `EarnConfirmationScreen` and its test file in
favour of the updated `EarnWithdrawConfirmationScreen` and its test
file.

### Test plan
After changing redirects to the new screen all the existing tests pass.

### Related issues

- Relates to
[ENG-223](https://linear.app/valora/issue/ENG-223/%5Bwallet%5D-add-new-reviewtransaction-component-to-earnconfirmationscreen)

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)