fix(shortcuts): error when claiming reward when extra divvi registration is needed (#198)
### Description

This fixes an error when claiming rewards from shortcuts, when an extra
divvi registration is also needed.

`prepareTransactions` was updated in
https://github.com/valora-inc/wallet/pull/6492 to insert a Divvi
registration when needed.
However `sendPreparedTransactions` filters out the Divvi registration
and then checks the `createBaseStandbyTransactions` array is the same
length as the filtered transactions. And throws if they don't match.

All call sites where updated to account for this "surprising" behavior,
except one that was missed, which handles shortcuts rewards claiming.

Note: we'll soon revamp this part with the new Divvi registration flow
where no extra TX will be needed. This will remove these "surprising"
behaviors. I omitted an extra unit test for this reason too.

### Test plan

- Manually tested claiming a reward works when an extra Divvi
registration is needed

### Related issues

- See Slack
[thread](https://valora-app.slack.com/archives/C04B61SJ6DS/p1744822525875159).

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)