fix(ci): ensure we scroll to the asset we tap (#212)
### Description

As the asset list on the test account grows, some items we tap on during
the E2E test may not be visible.

This ensures the item is scrolled into view before the tap.

Context: The asset list went bigger, because I added some ETH on
Optimism to burn unwanted spam NFTs that also blocked the CI. See also
#211

### Test plan

CI

### Related issues

NA

### Backwards compatibility

NA

### Network scalability

NA