fix: clear stored items on onboarding start (#208)
### Description

Some users are unable to proceed with onboarding due to the `Could not
decrypt data with alias: PEPPER` issue.

Sentry:
https://valora-inc.sentry.io/issues/?query=is%3Aunresolved%20issue.priority%3A%5Bhigh%2C%20medium%5D%20decrypt&referrer=issue-list&statsPeriod=14d

Slack:
https://valora-app.slack.com/archives/C04B61SJ6DS/p1744831162529559?thread_ts=1744827799.655419&cid=C04B61SJ6DS

This suggests that some stale keychain values are kept from the previous
installation. While the exact root cause is still unknown, this is
likely to happen after the app is reinstalled:
* The encryption key used for the previous app installation is gone
* The stored encrypted value remained for some reason: maybe it is
restored by the OS from backup on app reinstall for user convenience
alongside the other app data; maybe it is not properly wiped out

I wasn't able to reproduce it naturally in the Android emulator.

Our onboarding logic doesn't expect to have the stale keychain values.

This PR introduces the cleanup of keychain values when the user starts
the onboarding.

### Reproduction

In the Android emulator:

1) Install the app
2) In Android Studio Device Explorer navigate to `data/data/files/`
3) Create the directory `datastore`
4) Copy `RN_KEYCHAIN.preferences_pb` to this directory (download as
.zip:
[RN_KEYCHAIN.preferences_pb.zip](https://github.com/user-attachments/files/20012537/RN_KEYCHAIN.preferences_pb.zip))
-- this is the emulation of stale data, with no corresponding decryption
keys
5) Run the app and try to complete onboarding

<img width="661" alt="datastore"
src="https://github.com/user-attachments/assets/a17980a6-47ce-4eed-a272-3ca182d3c890"
/>

### Expected behavior

The stale data is ignored and the user is able to complete the
onboarding flow.


### Workaround

Delete app data using Android Settings

### Test plan

* Added unit test
* Tested manually

### Related issues

- Fixes ENG-300

### Backwards compatibility

Y

### Network scalability

NA