feat: upgrading from divvi v0 registration to divvi v1 registration (#210)
### Description

* Automatic divvi referrals using
[@divvi/referral-sdk](https://www.npmjs.com/package/@divvi/referral-sdk)
* Breaking change because referrerId and protocolId are no longer valid

### Test plan

<!-- Demonstrate the change is solid, or why it doesn't need testing.
Example: add any manual testing steps or scenarios (if not obvious),
screenshots / videos if the pull request changes the user interface.
-->

### Related issues


https://linear.app/valora/issue/ENG-299/update-divvi-framework-for-the-calldata-appending-referral-flow

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)