feat(home): Add Divvi bottom sheet to home screen (#216)
### Description

Fixes
[ENG-360](https://linear.app/valora/issue/ENG-360/divvi-slices-bottom-sheet)
and
[ENG-361](https://linear.app/valora/issue/ENG-361/divvi-slices-bottom-sheet-analytics-catch-all).

Adds a Divvi bottom sheet shown to the user a single time upon
navigating to the homescreen.

### Test plan

Manual and unit tested. See video.


https://github.com/user-attachments/assets/f91fd3ce-25ec-4cda-8636-d2489b22330d

### Related issues

- Fixes
[ENG-360](https://linear.app/valora/issue/ENG-360/divvi-slices-bottom-sheet)
and
[ENG-361](https://linear.app/valora/issue/ENG-361/divvi-slices-bottom-sheet-analytics-catch-all).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>