feat(divvi-protocol): update referral sdk to v2 (#223)
### Description

Update https://www.npmjs.com/package/@divvi/referral-sdk to v2

We had some fun redux state to manage which campaigns had already been
registered. With the new sdk that is no longer relevant, so most of the
code changes are around getting rid of that and references to
campaignIds

### Test plan

Unit tests updated. Verified with example app that referrals are
successful

### Related issues

https://linear.app/valora/issue/ENG-492/register-valora-app-with-divvi