chore(deps): replace react-native-fast-image with expo-image (#225)
### Description

As
[react-native-fast-image](https://github.com/DylanVann/react-native-fast-image/releases/tag/v8.6.3)
is no longer actively maintained and is blocking expo / react-native
upgrades, we've decided to replace it with
[expo-image](https://www.npmjs.com/package/expo-image).
`expo-image@2.0.7` was selected by using `npx expo install expo-image`,
which is the best way to find a compatible version with our current Expo
version.


### Test plan

- [x] Tested locally on Android & iOS by restoring a wallet with NFTs
and viewing the 'Collectable' tabs.
- [x] Unit Tests / Mocks Updated

### Related issues

ENG-499

### Backwards compatibility

Yes

### Network scalability

N/A