fix(divvi): divvi tag concatenation (#240)
### Description

When the Divvi tag is added to a `data` field, the undefined value is
stringified to a literal `"undefined"` value. The `data` is undefined in
the case of a native token transfer.

Example of a malformed request (transfer of 0.0001 ETH):

> ERROR Warning: src/send/usePrepareSendTransactions ::
prepareTransactionsOutput: EstimateGasExecutionError: Invalid parameters
were provided to the RPC method.
> Double check you have provided the correct parameters.
>
> URL: https://base-mainnet.g.alchemy.com/v2/
> Request body: {"method":"eth_estimateGas","params":
[{"data":"undefined6decb85d0100800000000000000000000000001ee8a7e8007fdd84fa050478be987cb8391cb9630000000000000000000000001ee8a7e8007fdd84fa050478be987cb8391cb96300000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000000000","from":"0x1ee8a7e8007fdd84fa050478be987cb8391cb963","maxFeePerGas":"0x2fa5ef","maxPriorityFeePerGas":"0xf4240","to":"0x1ee8a7e8007fdd84fa050478be987cb8391cb963","value":"0x5af3107a4000"}]}

This behavior makes it impossible to transfer ETH on Arbitrum, Base,
Ethereum, and Optimism.
CELO remains unaffected due to its dual nature, as it is transferred as
an ERC-20 token.

User feedback:
https://valora-app.slack.com/archives/C04B61SJ6DS/p1753974715455379

### Solution

The Divvi tag is NOT added to the `data` field if it is empty to avoid
compatibility issues: See [discussion
below](https://github.com/divvi-xyz/divvi-mobile/pull/240#discussion_r2332335009):

> We've had a couple of reports of native value transfers to smart
contracts reverting because the target doesn't have the selector. Since
our referral tag is not a real function.

| Issue repro | After fix |
|--------|--------|
| <img width="408" height="882" alt="Screenshot 2025-09-08 at 21 20 49"
src="https://github.com/user-attachments/assets/b1c94c19-d04b-458f-8484-afc35c30da1b"
/> | <img width="408" height="882" alt="Screenshot 2025-09-08 at 21 22
10"
src="https://github.com/user-attachments/assets/b1778379-ffc5-4aaa-96f5-e35d74dac928"
/> |

### Test plan

* Tested manually
* Updated unit test

### Related issues

NA

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)