fix: revert divvi-xyz/divvi-mobile#208 (#259)
### Description

https://github.com/divvi-xyz/divvi-mobile/pull/258 fixes the lower level
issue that was partially mitigated by
https://github.com/divvi-xyz/divvi-mobile/pull/208

I decided to revert it, because it's not necessary anymore and also
because it wipes everything in the keychain, even values that are set by
other libs (e.g. firebase items), which may or may not cause issues.

### Test plan

- Manually tested restoring an existing react-native-keychain datastore
after wiping the Keystore, still allows to restore or create a new
account. See https://github.com/divvi-xyz/divvi-mobile/pull/208
reproduction.

### Related issues

- Part of ENG-636

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)