chore(walletconnect): update walletConnect (#261)
### Description

Update walletConnect packages to the latest to facilitate smooth
EIP-5792 support implementation.

The current version doesn't allow specifying arbitrary
`sessionProperties`, which is important for a smooth UX when
communicating wallet capabilities to a dApp.

### Test plan

CI

### Related issues

- Related to ENG-644

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)