feat: add reCAPTCHA to phone verification and CAB flows (#233)
### Description

This PR adds recaptcha to the phone number verification and CAB restore
flows. There is no visible change for users, since the token is fetched
in the background based on the device and previous user interactions
with the app.
- the recaptcha client is a singleton
- it is instantiated on app start to begin collecting signals, but the
docs mention it can take seconds so it is not part of the blocking app
start actions
- the recaptcha tokens are very short living and it is recommended to
get them as late as possible. i've added them to the button presses,
unfortunately this will delay the phone number verification a little
- errors on fetching the recaptcha token are blocking on the phone
number verification flow but non blocking for the CAB flow

### Test plan

- Test locally on Android / iOS with local changes to the phone number
verification endpoint

### Related issues

- Part of ENG-442

### Backwards compatibility

Yes

### Network scalability

N/A

---------

Co-authored-by: Kathy Luo <kathyluo18@gmail.com>