feat(walletconnect): wallet_getCapabilities (#260)
### Description

Implement support for `wallet_getCapabilities` request (EIP-5792)
through WalletConnect.

EIP-5792:
* https://eips.ethereum.org/EIPS/eip-5792#wallet_getcapabilities
* https://www.eip5792.xyz/reference/getCapabilities

### Note on mobile UX

Whenever a dapp sends a WalletConnect request, the user must open their
wallet app to respond. The mobile app can’t process the request quietly
in the background. This means the user has to do an extra app switch,
which adds friction to the experience (see video below).

For instance, when using `useCapabilities` together with a WalletConnect
provider, as soon as the connection succeeds, `useCapabilities` will
automatically request the wallet’s capabilities. That triggers another
deep link into the wallet without any explanation to the user. From
their perspective, it might feel like the app is bouncing them around
for no clear reason, even like a bug.

To mitigate this, WalletConnect recommends sending the supported
capabilities within session data. They will be cached, and no extra
action from the user will be required. It will be addressed separately:
* https://github.com/divvi-xyz/divvi-mobile/pull/262


https://github.com/user-attachments/assets/473a6915-fc72-450c-8843-593a501718a5

### Test plan

* Tested manually using
https://bakoushin.github.io/wallet-connect-tester/
* Updated unit test

### Related issues

- Part of ENG-644

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)