feat(walletconnect): provide wallet capabilities within session data (#262)
### Description

Send wallet capabilities within the session data to avoid an extra
`wallet_getCapabilities` request that would otherwise require an extra
user interaction.

By including capabilities directly in the session, WalletConnect can
respond to `wallet_getCapabilities` without prompting the user to switch
to their wallet. This makes capability extraction seamless and improves
the overall UX.

### WalletConnect capabilities support

WalletConnect currently supports two approaches for including
capabilities in session data:

1. Recommended: in `sessionProperties` or `scopedProperties`.
[Docs](https://docs.walletconnect.network/wallet-sdk/react-native/eip5792#wallet-response)
2. Legacy: in the `capabilities` prop inside `sessionProperties`.
[Source](https://github.com/WalletConnect/walletconnect-monorepo/blob/1e6d7793d0a30d2bf684cd3811ba120b4cdc0498/providers/universal-provider/src/providers/eip155.ts#L219-L223)

We do support both approaches for maximal compatibility.

### Notes

* When using the recommended approach, WalletConnect does not return
capabilities if the optional `chainIds` array is omitted in
`wallet_getCapabilities` request.
* We've proposed a PR to address this in the WalletConnect repo:
https://github.com/WalletConnect/walletconnect-monorepo/pull/6955

### Test plan

* Updated unit test
* Tested manually using
https://bakoushin.github.io/wallet-connect-tester/


https://github.com/user-attachments/assets/bd3a6d15-4dd1-4dff-a295-a49828500d45

### Related issues

- Part of ENG-644

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)