chore(walletconnect): refactor to support many txs (#267)
### Description

Refactor WalletConnect codepath to handle multiple transactions (instead
of a single one).

This change lays the groundwork for the upcoming `wallet_sendCalls`
method implementation, where multiple transactions need to be executed
sequentially when atomic execution is neither requested by the dapp nor
supported by the wallet.

### Review notes

Better reviewed with collapsed whitespace.

### Related work

* https://github.com/divvi-xyz/divvi-mobile/pull/268

### Test plan

* CI
* Updated unit tests

### Related issues

- Related to ENG-644

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)