fix(celo): Catch gas estimation error when low CELO balance (#269)
### Description

When a user had a very low CELO balance (~0.00001), send/swap/etc. would
fail on Celo. This is because this [error
handling](https://github.com/divvi-xyz/divvi-mobile/blob/2bf6b24f85d136926c36d77a9976a7aa1c107ca0/packages/%40divvi/mobile/src/viem/prepareTransactions.ts#L181-L196)
was failing to catch the error, it had the right name but wasn't being
caught in `instanceof`.

### Test plan

Created a new wallet with 1 cUSD and 0.00001 CELO. Before fix if I tried
to send 0.5 cUSD the prepare transaction logic would fail with a generic
failure error. After the fix, the transaction prepares successfully with
cUSD as fee token.

### Related issues

- Fixes ENG-662

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)