chore(walletconnect): refine typings (#270)
### Description

Refine typings to reduce ambiguity and improve type safety. Follow-up to
https://github.com/divvi-xyz/divvi-mobile/pull/267#discussion_r2402710087.

### Key types

* `ActionableRequest` type that has explicitly typed `method`. It
enables proper type discrimination based on the method. This type is
passed down the line to the `handleRequest` method, through
`ActionRequest` component. It is basically a wrapper around the standard
WC `session_request` type with a strongly typed method.
* `PreparedTransaction` type is discriminated by the `success` property,
ensuring that either `transactionRequest` or `errorMessage` to be
present, instead of having both optional.

### Test plan

* CI
* Updated unit tests

### Related issues

- Related to ENG-644

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)