chore(walletconnect): check `wallet_getCapabilities` validity before handling request (#273)
### Description

Currently we are checking `wallet_getCapabilities` parameters validity
while executing the request. This is too late, because at this point we
normally assume that all checks have been complete.
We should proceed with all checks before executing the request, and
answer to the client with a proper error code, specified in the
[EIP-5792](https://eips.ethereum.org/EIPS/eip-5792#error-codes).

This moves checks to the `showActionRequest` saga.

### Test plan

* Updated unit test

### Related issues

- Related to ENG-644

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)