fix: fee calculation error matching (#277)
### Description

Follow up to #269 - we should probably also apply the change to the
conditions for `cause` so that the fee estimation logic runs as
intended. The main outcome is that the correct / more informative "you
need more CELO for gas" warning is displayed instead of the generic
"failed to prepare a transaction" when a user tries to create a
transaction that is too close to their balance amount.

Also included are some yarn.lock changes, I went on a journey to try to
make `instanceof` work properly and thought that having 2 versions of
`viem` might be the culprit. It didn't fix the bug, but having 2
versions is unnecessary so I want to check in the changes anyway.

### Test plan

Tested that a user with very low CELO balance + only cUSD + sending
close to the cUSD balance results in the correct low balance warning.

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)