chore(walletconnect): `wallet_sendCalls` requested capabilities validation (#276)
### Description

A dApp may request specific capabilities within the `walled_sendCalls`
request.

[According to the
spec](https://eips.ethereum.org/EIPS/eip-5792#wallet_sendcalls), if some
of the requested capabilities are not supported by the wallet, and are
not marked as optional in the request, the request must be rejected.

> The wallet:
> ...
> * MUST reject the request if it contains a capability (either
top-level or call-level) that is not supported by the wallet and the
capability is not explicitly marked as optional.
> * Applications may mark a capability as optional by setting optional
to true. See the RPC Specification section below for details.

The validation is not hidden under any feature flag, since it is not
expected to cause any harm (the request will be rejected anyways).

### Test plan

* Added unit test
* Tested manually against
https://bakoushin.github.io/wallet-connect-tester/

| Atomic is required | Atomic is optional |
|--------|--------|
| Request denied with "Atomicity not supported" | Request denied with
"Unsupported method" (for now) |
| <img width="1320" height="2868" alt="image"
src="https://github.com/user-attachments/assets/97f56f4b-6088-4c26-9476-04eefa713507"
/> | <img width="1320" height="2868" alt="image"
src="https://github.com/user-attachments/assets/97ea3954-ffe4-4e58-852f-0e36113966f5"
/> |

### Related issues

- Part of ENG-644

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)